requirements.txt
numpy==2.2.6
torch==2.7.1
torchvision==0.22.1
faiss-cpu==1.11.0
Pillow==11.3.0
tqdm==4.67.1
regex==2024.11.6
ftfy==6.3.1
clip==1.0
